Task 3 - AI Resume Screening System with Tracing/ai_resume_screener/requirements.txt
langchain
langchain-core
langchain-groq
python-dotenv
langsmith